backend/think-server/requirements.txt
fastapi
uvicorn
langchain-google-genai
google-genai
google-generativeai
